// web-application/src/types-openapi/api/models/UpdateVoucherRequest.ts
export interface UpdateVoucherRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateVoucherRequest
     */
    code?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateVoucherRequest
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateVoucherRequest
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateVoucherRequest
     */
    discountType?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVoucherRequest
     */
    discountValue?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateVoucherRequest
     */
    discountPercentage?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVoucherRequest
     */
    maxDiscountValue?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVoucherRequest
     */
    minOrderValue?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof UpdateVoucherRequest
     */
    startAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof UpdateVoucherRequest
     */
    endAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof UpdateVoucherRequest
     */
    usageLimitTotal?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateVoucherRequest
     */
    usageLimitPerUser?: number;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateVoucherRequest
     */
    isActive?: boolean;
    /**
     * 
     * @type {Array<VoucherTimeRuleDto>}
     * @memberof UpdateVoucherRequest
     */
    timeRules?: Array<VoucherTimeRuleDto> | null;
    /**
     * 
     * @type {Array<VoucherUserRuleDto>}
     * @memberof UpdateVoucherRequest
     */
    userRules?: Array<VoucherUserRuleDto> | null;
}